design: 핀 뽑아오는 컴포넌트명 수정 및 fixed 부여
